Add reservation form and confirmation components; implement availability check and newsletter opt-in
